the setup using vite for testing and future modules is ready
I used vite to start a project inside the repo so that I can use It's capibilites and also vite is very well suited for working for TS I have realized. I had not decided on a Front End frame work yet but I'm most familliar with React so I decided to go with it. But that's not important right now the data for the app is not yet complete